pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fb-groups-poster"
version = "0.1.0"
description = "Automated Facebook groups poster with Google Sheets logging"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Abdelrahman Habib"}]
dependencies = [
  "click>=8.2.1",
  "PyYAML>=6.0.2",
  "gspread>=6.2.1",
  "google-auth>=2.40.3",
  "selenium>=4.34.2",
  "webdriver-manager>=4.0.2",
  "pyperclip>=1.9"
]

[project.scripts]
fbpost = "fb_groups_poster.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
